Givens orthogonalisation

Orthogonalization using Givens' method.

givens_orth(X, nullspace = FALSE)

Arguments

X
a numeric matrix with ncol(X) <= nrow(X)

nullspace
logical: do you want an orthogonal basis for the null space?

Value

A list with components Q, R, as normally defined, and if nullspace is TRUE a further component N giving the basis for the requested null space of X

Examples

set.seed(1234)
X <- matrix(rnorm(7*6), 7)
givens_orth(X, nullspace = TRUE)

#> $Q
#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#> [1,] -0.40273912 -0.31886616  0.35216372 -0.08746729  0.62289875 -0.4564106
#> [2,]  0.09256464 -0.31031425 -0.08537802 -0.13665346 -0.54365576 -0.5419407
#> [3,]  0.36182526 -0.48126459 -0.28007387  0.46655507  0.06876516 -0.2936426
#> [4,] -0.78264520 -0.29442411 -0.24173430  0.30962958 -0.29082899  0.2451163
#> [5,]  0.14317803 -0.54964063 -0.37088570 -0.54560652  0.23277688  0.4171142
#> [6,]  0.16884623 -0.42520658  0.76713975  0.08795868 -0.26909923  0.3448704
#> [7,] -0.19176276  0.02868483  0.07586076 -0.59557366 -0.31705825 -0.2424083
#> 
#> $R
#>          [,1]        [,2]       [,3]      [,4]       [,5]      [,6]
#> [1,] 2.997141 -0.06703921  0.3939392 0.9520604  0.8167085 0.3965354
#> [2,] 0.000000  1.79897163 -0.3207575 0.7984742  0.4611396 1.3899184
#> [3,] 0.000000  0.00000000  2.8846528 0.8041695 -0.3641774 1.4253453
#> [4,] 0.000000  0.00000000  0.0000000 1.5531160  1.8094798 0.7014172
#> [5,] 0.000000  0.00000000  0.0000000 0.0000000  1.1997846 0.2915716
#> [6,] 0.000000  0.00000000  0.0000000 0.0000000  0.0000000 2.6006064
#> 
#> $N
#>             [,1]
#> [1,]  0.09023434
#> [2,]  0.52906706
#> [3,] -0.50039706
#> [4,] -0.04256158
#> [5,]  0.11825078
#> [6,] -0.05567945
#> [7,] -0.66532222
#>