Naive backeward elimination

A simple facility to refine models by backward elimination. Covers cases where drop_term works but step_AIC does not

step_down(object, ..., trace = FALSE, k)

Arguments

object
A fitted model object

...
additional arguments passed to drop_term such as k

trace
logical: do you want a trace of the process printed?

k
penalty (default 2, as for AIC)

Value

A refined fitted model object

Examples

fm <- lm(medv ~ . + (rm + tax + lstat)^2 +
           I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 12)^2), Boston)
sfm <- step_down(fm, trace = TRUE, k = "bic")

#> Single term deletions
#> 
#> Model:
#> medv ~ crim + zn + indus + chas + nox + rm + age + dis + rad + 
#>     tax + ptratio + black + lstat + (rm + tax + lstat)^2 + I((rm - 
#>     6)^2) + I((tax - 400)^2) + I((lstat - 12)^2)
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> indus              1      0.22 6044.8 1373.4   0.017 0.8950975    
#> age                1     21.25 6065.8 1375.2   1.709 0.1917385    
#> rm:lstat           1     22.43 6067.0 1375.2   1.804 0.1799023    
#> zn                 1     23.04 6067.6 1375.3   1.853 0.1740887    
#> black              1     42.22 6086.8 1376.9   3.395 0.0660202 .  
#> I((rm - 6)^2)      1     55.87 6100.4 1378.0   4.492 0.0345550 *  
#> I((tax - 400)^2)   1     57.16 6101.7 1378.1   4.596 0.0325442 *  
#> <none>                         6044.5 1379.6                      
#> chas               1     89.99 6134.5 1380.8   7.236 0.0073935 ** 
#> rad                1    141.11 6185.7 1385.0  11.345 0.0008164 ***
#> nox                1    205.38 6249.9 1390.3  16.513 5.633e-05 ***
#> crim               1    261.65 6306.2 1394.8  21.038 5.737e-06 ***
#> ptratio            1    429.13 6473.7 1408.1  34.503 7.901e-09 ***
#> I((lstat - 12)^2)  1    497.63 6542.2 1413.4  40.011 5.734e-10 ***
#> dis                1    544.79 6589.3 1417.0  43.803 9.619e-11 ***
#> rm:tax             1    862.92 6907.5 1440.9  69.382 8.335e-16 ***
#> tax:lstat          1   1443.12 7487.7 1481.7 116.031 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  indus 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + zn + chas + nox + rm + age + dis + rad + tax + 
#>     ptratio + black + lstat + I((rm - 6)^2) + I((tax - 400)^2) + 
#>     I((lstat - 12)^2) + rm:tax + rm:lstat + tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> age                1     21.31 6066.1 1369.0   1.717 0.1907477    
#> rm:lstat           1     22.81 6067.6 1369.1   1.838 0.1758125    
#> zn                 1     22.83 6067.6 1369.1   1.839 0.1756806    
#> black              1     42.01 6086.8 1370.7   3.384 0.0664241 .  
#> I((rm - 6)^2)      1     55.66 6100.4 1371.8   4.484 0.0347136 *  
#> I((tax - 400)^2)   1     63.98 6108.7 1372.5   5.155 0.0236184 *  
#> <none>                         6044.8 1373.4                      
#> chas               1     90.82 6135.6 1374.7   7.317 0.0070713 ** 
#> rad                1    156.70 6201.5 1380.1  12.625 0.0004178 ***
#> nox                1    218.98 6263.7 1385.2  17.643 3.171e-05 ***
#> crim               1    261.67 6306.4 1388.6  21.081 5.610e-06 ***
#> ptratio            1    430.48 6475.2 1402.0  34.682 7.243e-09 ***
#> I((lstat - 12)^2)  1    498.55 6543.3 1407.3  40.166 5.320e-10 ***
#> dis                1    576.55 6621.3 1413.3  46.450 2.786e-11 ***
#> rm:tax             1    866.77 6911.5 1435.0  69.832 6.796e-16 ***
#> tax:lstat          1   1491.21 7536.0 1478.7 120.141 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  age 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + zn + chas + nox + rm + dis + rad + tax + ptratio + 
#>     black + lstat + I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 
#>     12)^2) + rm:tax + rm:lstat + tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> rm:lstat           1     22.45 6088.5 1364.6   1.806 0.1796030    
#> zn                 1     27.36 6093.4 1365.0   2.201 0.1385832    
#> black              1     39.72 6105.8 1366.0   3.195 0.0744726 .  
#> I((rm - 6)^2)      1     55.94 6122.0 1367.4   4.500 0.0343891 *  
#> I((tax - 400)^2)   1     64.55 6130.6 1368.1   5.193 0.0231129 *  
#> <none>                         6066.1 1369.0                      
#> chas               1     87.56 6153.6 1370.0   7.044 0.0082127 ** 
#> rad                1    163.30 6229.4 1376.2  13.137 0.0003199 ***
#> crim               1    271.26 6337.3 1384.8  21.822 3.874e-06 ***
#> nox                1    277.53 6343.6 1385.4  22.327 3.012e-06 ***
#> ptratio            1    452.85 6518.9 1399.2  36.430 3.134e-09 ***
#> I((lstat - 12)^2)  1    540.12 6606.2 1405.9  43.451 1.130e-10 ***
#> dis                1    560.05 6626.1 1407.4  45.055 5.334e-11 ***
#> rm:tax             1    846.58 6912.7 1428.8  68.106 1.457e-15 ***
#> tax:lstat          1   1481.71 7547.8 1473.3 119.200 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  rm:lstat 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + zn + chas + nox + rm + dis + rad + tax + ptratio + 
#>     black + lstat + I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 
#>     12)^2) + rm:tax + tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> zn                 1     25.42 6113.9 1360.5   2.041 0.1537061    
#> black              1     56.56 6145.1 1363.0   4.542 0.0335620 *  
#> I((tax - 400)^2)   1     60.73 6149.3 1363.4   4.878 0.0276663 *  
#> <none>                         6088.5 1364.6                      
#> chas               1     97.84 6186.4 1366.4   7.858 0.0052606 ** 
#> I((rm - 6)^2)      1    118.65 6207.2 1368.1   9.529 0.0021370 ** 
#> rad                1    153.81 6242.3 1371.0  12.353 0.0004812 ***
#> crim               1    264.05 6352.6 1379.8  21.207 5.263e-06 ***
#> nox                1    284.53 6373.1 1381.5  22.852 2.318e-06 ***
#> ptratio            1    462.96 6551.5 1395.4  37.183 2.186e-09 ***
#> dis                1    559.65 6648.2 1402.9  44.949 5.594e-11 ***
#> rm:tax             1    957.18 7045.7 1432.2  76.876 < 2.2e-16 ***
#> I((lstat - 12)^2)  1   1146.09 7234.6 1445.6  92.048 < 2.2e-16 ***
#> tax:lstat          1   1481.42 7569.9 1468.6 118.980 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  zn 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + chas + nox + rm + dis + rad + tax + ptratio + black + 
#>     lstat + I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 12)^2) + 
#>     rm:tax + tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> black              1     56.25 6170.2 1358.9   4.508  0.034237 *  
#> I((tax - 400)^2)   1     68.49 6182.4 1359.9   5.489  0.019536 *  
#> <none>                         6113.9 1360.5                      
#> chas               1     99.42 6213.4 1362.4   7.968  0.004954 ** 
#> I((rm - 6)^2)      1    113.05 6227.0 1363.5   9.060  0.002747 ** 
#> rad                1    140.54 6254.5 1365.7  11.263  0.000852 ***
#> crim               1    256.32 6370.3 1375.0  20.543 7.337e-06 ***
#> nox                1    294.29 6408.2 1378.0  23.586 1.610e-06 ***
#> ptratio            1    585.16 6699.1 1400.5  46.897 2.249e-11 ***
#> dis                1    606.13 6720.1 1402.1  48.578 1.030e-11 ***
#> rm:tax             1   1004.67 7118.6 1431.2  80.519 < 2.2e-16 ***
#> I((lstat - 12)^2)  1   1196.66 7310.6 1444.7  95.906 < 2.2e-16 ***
#> tax:lstat          1   1475.51 7589.4 1463.6 118.255 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  black 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + chas + nox + rm + dis + rad + tax + ptratio + lstat + 
#>     I((rm - 6)^2) + I((tax - 400)^2) + I((lstat - 12)^2) + rm:tax + 
#>     tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> I((tax - 400)^2)   1     71.80 6242.0 1358.5   5.713  0.017214 *  
#> <none>                         6170.2 1358.9                      
#> chas               1    104.89 6275.1 1361.2   8.347  0.004035 ** 
#> I((rm - 6)^2)      1    113.40 6283.6 1361.9   9.024  0.002801 ** 
#> rad                1    129.12 6299.3 1363.1  10.275  0.001436 ** 
#> crim               1    284.63 6454.8 1375.5  22.650 2.560e-06 ***
#> nox                1    311.95 6482.1 1377.6  24.824 8.718e-07 ***
#> ptratio            1    567.04 6737.2 1397.1  45.123 5.135e-11 ***
#> dis                1    612.89 6783.1 1400.6  48.772 9.389e-12 ***
#> rm:tax             1   1044.43 7214.6 1431.8  83.112 < 2.2e-16 ***
#> I((lstat - 12)^2)  1   1228.20 7398.4 1444.5  97.735 < 2.2e-16 ***
#> tax:lstat          1   1562.29 7732.5 1466.9 124.321 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Eliminating:  I((tax - 400)^2) 
#> -----------
#> Single term deletions
#> 
#> Model:
#> medv ~ crim + chas + nox + rm + dis + rad + tax + ptratio + lstat + 
#>     I((rm - 6)^2) + I((lstat - 12)^2) + rm:tax + tax:lstat
#>                   Df Sum of Sq    RSS    BIC F Value     Pr(F)    
#> <none>                         6242.0 1358.5                      
#> I((rm - 6)^2)      1    105.66 6347.6 1360.8   8.329  0.004074 ** 
#> chas               1    112.78 6354.8 1361.3   8.889  0.003010 ** 
#> crim               1    291.81 6533.8 1375.4  23.001 2.150e-06 ***
#> nox                1    338.73 6580.7 1379.0  26.699 3.462e-07 ***
#> rad                1    369.65 6611.6 1381.4  29.136 1.051e-07 ***
#> ptratio            1    520.53 6762.5 1392.8  41.029 3.511e-10 ***
#> dis                1    660.61 6902.6 1403.2  52.070 2.040e-12 ***
#> rm:tax             1   1128.07 7370.1 1436.3  88.916 < 2.2e-16 ***
#> I((lstat - 12)^2)  1   1194.41 7436.4 1440.9  94.145 < 2.2e-16 ***
#> tax:lstat          1   1496.15 7738.1 1461.0 117.928 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1